requirements.txt
streamlit
google-generativeai
faiss-cpu
pypdf
python-dotenv
numpy